add more functionality to vertical toolbar
The original vertical toolbar was static (not responsive). Functionality was added to make the toolbar flip it's orientation from vertical to horizontal while being displayed in portrait mode on smaller devices.

Since the vertical toolbar is contained within a module content area such as dashboard.component, some styles had to be added to the module content styles sheets to get the toolbar to style correctly.