Add a sidenav to house the main navigation
A sidenav was added to provide navigation for mobile devices. The main module buttons disappear and are replaced by a menu button that reveals the new sidenav.

Headers were also added to the tool sections so that users can know where in the app they are.

Router links were also added to all buttons that currently have somewhere to navigate to

RouterLinkActive was added to all the appropriate buttons. This makes it so the "active" CSS class gets added to link elements when it's route is active.

Some fixes were made to the app theme. In particular, when spot treating certain elements, such as an md-icon, if the element was nested within an app theme, the theme applied to the element theme would be overwritten by the theme found higher in the heirarchy. This had to do with the order in work angular material applied custom theme to material componenets. A fix was made until a more permanent solution can be made.