Add multiple module sections and components
Added the main modules and cooresponding navigation. You can now click on the various items in the vertical toolbars of each module and navigate to their respective sections

A first attempt of a notifications sidenav was created. Md-tabs proved to be difficult to work with so some temporary custom styles had to be made to get the tabs to have the proper width.

Workspace routes were moved from workspace-routing.module.ts to their own .routing.ts file