add account feature area and base services
The account area was added to give the account owner/administrator a way to configure various account settings. At the moment it is just a shell. But the base layout is complete.

Another menu button was added to the main workspace toolbar. It currently has one button which is to navigate to the account area of the app. A temorary button was also added to the account sidenav so that you can navigate back to dashboard. The account area is separate from the workspace and is lazy loaded in since most users won't be accessing this part of the app.

Various common services were also added. These do not have any implementation as of yet but that will be added as things move foreward.